requirements.txt
torch
transformers
scikit-learn
numpy
joblib
rapidfuzz
tldextract
pandas
httpx
urllib3
flask
xgboost
psycopg2-binary
